+++ base-requirements.txt
@@ -23,6 +23,7 @@ idna==2.1
 ipaddress==1.0.16
 jmespath==0.9.0
 marshmallow==2.10.2
+marshmallow_polyfield==3.1
 pyOpenSSL==16.1.0
 pyasn1==0.1.9
 pyasn1-modules==0.0.8
